# docs/user-api-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "B3Cloud User API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/deployments": {
            "get": {
                "summary": "V1 List Deployments",
                "operationId": "v1_list_deployments_api_v1_deployments_get",
                "parameters": [
                    {
                        "name": "namespace",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Namespace"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "additionalProperties": true
                                    },
                                    "title": "Response V1 List Deployments Api V1 Deployments Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "V1 Create Deployment",
                "operationId": "v1_create_deployment_api_v1_deployments_post",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AppDeployIn"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response V1 Create Deployment Api V1 Deployments Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AppDeployIn": {
                "properties": {
                    "github_url": {
                        "type": "string",
                        "title": "Github Url"
                    },
                    "env": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Env"
                    },
                    "git_revision": {
                        "type": "string",
                        "title": "Git Revision",
                        "default": "main"
                    },
                    "port": {
                        "type": "integer",
                        "title": "Port",
                        "default": 8080
                    },
                    "node_arch": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Node Arch"
                    },
                    "auto_detect_services": {
                        "type": "boolean",
                        "title": "Auto Detect Services",
                        "default": true
                    },
                    "provision_services": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Provision Services"
                    },
                    "redeploy_services": {
                        "type": "boolean",
                        "title": "Redeploy Services",
                        "default": false
                    },
                    "components": {
                        "items": {
                            "$ref": "#/components/schemas/ComponentDeployIn"
                        },
                        "type": "array",
                        "title": "Components"
                    },
                    "resources": {
                        "$ref": "#/components/schemas/ResourceLimitsIn"
                    },
                    "autoscaling": {
                        "$ref": "#/components/schemas/AutoscalingIn"
                    }
                },
                "type": "object",
                "required": [
                    "github_url"
                ],
                "title": "AppDeployIn"
            },
            "AutoscalingIn": {
                "properties": {
                    "enabled": {
                        "type": "boolean",
                        "title": "Enabled",
                        "default": true
                    },
                    "min_replicas": {
                        "type": "integer",
                        "maximum": 100.0,
                        "minimum": 1.0,
                        "title": "Min Replicas",
                        "default": 1
                    },
                    "max_replicas": {
                        "type": "integer",
                        "maximum": 100.0,
                        "minimum": 1.0,
                        "title": "Max Replicas",
                        "default": 5
                    },
                    "target_cpu_utilization": {
                        "type": "integer",
                        "maximum": 100.0,
                        "minimum": 1.0,
                        "title": "Target Cpu Utilization",
                        "default": 80
                    },
                    "target_memory_utilization": {
                        "type": "integer",
                        "maximum": 100.0,
                        "minimum": 1.0,
                        "title": "Target Memory Utilization",
                        "default": 80
                    }
                },
                "type": "object",
                "title": "AutoscalingIn"
            },
            "ComponentDeployIn": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "path": {
                        "type": "string",
                        "title": "Path",
                        "default": "."
                    },
                    "type": {
                        "type": "string",
                        "title": "Type",
                        "default": "backend"
                    },
                    "public": {
                        "type": "boolean",
                        "title": "Public",
                        "default": true
                    },
                    "port": {
                        "type": "integer",
                        "title": "Port",
                        "default": 8080
                    },
                    "auto_detect_services": {
                        "type": "boolean",
                        "title": "Auto Detect Services",
                        "default": true
                    },
                    "provision_services": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Provision Services"
                    },
                    "redeploy_services": {
                        "type": "boolean",
                        "title": "Redeploy Services",
                        "default": false
                    },
                    "autoscaling": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/AutoscalingIn"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "env": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Env"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "ComponentDeployIn"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ResourceLimitsIn": {
                "properties": {
                    "cpu_request": {
                        "type": "string",
                        "title": "Cpu Request",
                        "default": "100m"
                    },
                    "cpu_limit": {
                        "type": "string",
                        "title": "Cpu Limit",
                        "default": "500m"
                    },
                    "memory_request": {
                        "type": "string",
                        "title": "Memory Request",
                        "default": "128Mi"
                    },
                    "memory_limit": {
                        "type": "string",
                        "title": "Memory Limit",
                        "default": "512Mi"
                    }
                },
                "type": "object",
                "title": "ResourceLimitsIn"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}